# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AddCommentRequest:
    properties:
      body:
        maxLength: 150
        type: string
      receiver-id:
        type: string
      user-id:
        type: string
      whom:
        maxLength: 6
        type: string
    required:
    - body
    - receiver-id
    - user-id
    - whom
    type: object
...
  dto.LoginRequest:
    properties:
      login:
        maxLength: 100
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
...
  dto.RegisterRequest:
    properties:
      login:
        maxLength: 100
        type: string
      password:
        type: string
      telegram:
        type: string
    required:
    - login
    - password
    - telegram
    type: object
...
  entities.Comment:
    properties:
      author_id:
        type: string
      body:
        type: string
      comment_id:
        type: string
      time:
        type: string
    type: object
...
host: localhost:1111
info:
  contact: {}
  description: Ненавижу эту хуйню бля
  title: На жри уебок доки свои не подавись (недоделанная кстати хихиххихихих))))))
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: User login and receiving JWT token
      parameters:
      - description: Login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      summary: User authentication
      tags:
      - auth
  /auth/logout:
    post:
      description: Clears the JWT cookie
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /auth/profile:
    get:
      description: Returns the profile data of the current user
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Getting a logged in profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user in the system
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      summary: User registration
      tags:
      - auth
  /comments:
    get:
      description: Returns a list of comments by parameters
      parameters:
      - description: whose
        in: query
        name: whose
        type: string
      - description: user ID
        in: query
        name: userId
        type: string
      - description: news ID
        in: query
        name: newsId
        type: string
      - description: amount
        in: query
        name: amount
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of comments
          schema:
            items:
              $ref: '#/definitions/entities.Comment'
            type: array
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      summary: Receiving comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Creates a new comment
      parameters:
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Comment'
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Adding a comment
      tags:
      - comments
...
swagger: "2.0"